# generated/sonar-coverage-server/model/model_coverage.go
/*
 * Sonar coverage dashboard
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package model

import (
	"time"
)

type Coverage struct {
	Date time.Time `json:"date,omitempty"`

	Products []ProductCoverage `json:"products,omitempty"`
	// Coverage status
	Status string `json:"status,omitempty"`

	Errors []string `json:"errors,omitempty"`
}