# generated/sonar-coverage-server/model/model_product.go
/*
 * Sonar coverage dashboard
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package model

type Product struct {
	Name string `json:"name,omitempty"`

	Repos []ProductRepos `json:"repos,omitempty"`
}