# generated/sonar-coverage-server/model/model_product_coverage.go
/*
 * Sonar coverage dashboard
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package model

type ProductCoverage struct {
	LinesToCoverTotal int32 `json:"linesToCoverTotal,omitempty"`

	UnitTestsTotal int32 `json:"unitTestsTotal,omitempty"`

	CoverageTotal float64 `json:"coverageTotal,omitempty"`

	Repos []RepoCoverage `json:"repos,omitempty"`

	Name string `json:"name,omitempty"`
}