# logic/api_coverage.go
/*
 * Sonar coverage dashboard
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package logic

import (
	"net/http"

	"github.com/moznobkin/sonar-coverage/generated/sonar-coverage-server/model"
)

func GetCoverage(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	var products []model.Product
	err := json.NewDecoder(r.Body).Decode(&products)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	coverage, err := AggregateMetrics(r.Context(), products)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		return
	}
	json.NewEncoder(w).Encode(coverage)

	w.WriteHeader(http.StatusOK)
}

func GetDefaultCoverage(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	products, err := ReadStruct[[]model.Product]("./data/json/products.json")
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		return
	}

	coverage, err := AggregateMetrics(r.Context(), *products)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		return
	}
	json.NewEncoder(w).Encode(coverage)

	w.WriteHeader(http.StatusOK)
}